+++ src/decorators/index.ts
@@ -1,0 +1,5 @@
+export { AutoRegister } from "./autoRegister";
+export { Inject } from "./inject";
+export { Injectable } from "./injectable";
+export { InjectAll } from "./injectAll";
+export { Scoped } from "./scoped";
