+++ src/decorators/scoped.ts
@@ -1,0 +1,31 @@
+import { getMetadata } from "../metadata";
+import type { Scope } from "../scope";
+import type { Constructor } from "../token";
+import type { ClassDecorator } from "./decorators";
+
+/**
+ * Decorator for setting the scope of a class when registering.
+ *
+ * @example
+ * ```ts
+ * @Scoped(Scope.Container)
+ * class Wizard {}
+ *
+ * container.register(Wizard);
+ *
+ * // Under the hood
+ * container.register(
+ *   Wizard,
+ *   {useClass: Wizard},
+ *   {scope: Scope.Container},
+ * );
+ * ```
+ *
+ * @__NO_SIDE_EFFECTS__
+ */
+export function Scoped<This extends object>(scope: Scope): ClassDecorator<Constructor<This>> {
+  return (Class) => {
+    const metadata = getMetadata(Class);
+    metadata.scope = scope;
+  };
+}
